Session::request() : type + approprié pour les entiers.
La plupart des variables entières de Session::request() comptent un nombre de réponses, or il ne peut pas y avoir plus de 254 joueurs (les IDs 254 et 255 étant réservés), il est alors plus cohérent de stocker ces variables sur un byte, comme les IDs des joueurs.

+++ include/Rbo/ReplyHandler.hpp
@@ -17,8 +17,8 @@ struct RequestProfile {
 struct RequestCtx {
     std::map<byte, RequestProfile> players;
     Replies replies;
-    std::atomic<ulong> repliesHandled;
-    std::atomic<ulong> repliesToAccept;
+    std::atomic<byte> repliesHandled;
+    std::atomic<byte> repliesToAccept;
     std::atomic_bool requestDone;
     std::vector<byte> errorIDs;
